requirements.txt
streamlit>=1.20.0
numpy>=1.20.0
Pillow>=8.0.0
opencv-python>=4.5.0